Don't get fly if it already exists

+++ fly/fly.py
@@ -17,14 +17,15 @@ class Fly:
         self.target = target
 
     def get_fly(self):
-        url = f'{self.concourse_url}/api/v1/cli?arch=amd64&platform=' \
-            f'{self.platform}'
-        response = requests.get(url, stream=True)
-        if response.status_code == 200:
-            with open(self.executable, 'wb') as f:
-                for chunk in response:
-                    f.write(chunk)
-            os.chmod(self.executable, 0o755)
+        if not os.path.isfile(self.executable):
+            url = f'{self.concourse_url}/api/v1/cli?arch=amd64&platform=' \
+                f'{self.platform}'
+            response = requests.get(url, stream=True)
+            if response.status_code == 200:
+                with open(self.executable, 'wb') as f:
+                    for chunk in response:
+                        f.write(chunk)
+                os.chmod(self.executable, 0o755)
 
     def run(self, cmd, *args):
         return subprocess.run(
